backend/scraper/requirements.txt
playwright==1.41.0
beautifulsoup4==4.12.3
httpx==0.27.0
pika==1.3.2
redis==5.0.1
lxml==5.1.0
python-dotenv==1.0.1
